% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/category_mat.R
\name{category.mat}
\alias{category.mat}
\title{Converting category names into matrices and matrices int category names}
\usage{
category.mat(input = "Immediate")
}
\arguments{
\item{input}{}
}
\value{
list
}
\description{
Converting category names into matrices and matrices int category names
}
\examples{
\dontrun{category.mat(list("Immediate", matrix(c(0,0,0,0,0,0,0,0,0), nrow = 3)))}
}
