% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transriv_preprocessing.R
\name{combine_csv_mat}
\alias{combine_csv_mat}
\title{Combining the csv and mat data into csv file}
\usage{
combine_csv_mat(csv_path, mat_path)
}
\arguments{
\item{csv_path}{character -- the fullpath of csv file}

\item{mat_path}{character -- the fullpath of mat file}
}
\value{
data.frame
}
\description{
Combining the csv and mat data into csv file
}
\note{
This function uses \link{read_csv_data} and \link{read_mat_data}
    functions while reading the files.
}
\examples{
\dontrun{
combine_csv_mat(csv_path,mat_path)}
}
