% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_demog_trans.R
\name{combine_demog_trans}
\alias{combine_demog_trans}
\title{Combining the demographics and transition questionnaires}
\usage{
combine_demog_trans(csv_path, mat_path)
}
\arguments{
\item{mat_path}{character -- the fullpath of mat files}
}
\value{
data.frame
}
\description{
This function combines the demographics and the quantitative data of
transition questionnaire.
}
\note{
This function uses \link{read_csv_data} and \link{read_mat_data}
    functions while reading the files.

Use this function only for one subject at a time!
}
\examples{
\dontrun{
curr_subj <- combine_csv_mat(mat_path)
all_subj  <- rbind(all_subj,curr_subj)
all_data  <- cbind(data, all_subj)}
}
