% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transriv_statistics.R
\name{convert_speed}
\alias{convert_speed}
\title{Convert the speed values}
\usage{
convert_speed(data)
}
\arguments{
\item{data}{data.frame -- preprocessed transition data}
}
\value{
data.frame -- transition data with corrected speed values
}
\description{
The speed parameter was collected in just an opposite way during the
experiment. The maximum speed was recorded as 0 and the minimum was 100. This
function corrects the speed values.
}
\examples{
\dontrun{
data <- read_csv_data(csv_path)
data <- predefine_immediate(data,0)
data <- convert_speed(data)}
}
