% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_absence_presence_matrix.R
\name{create_absence_presence_matrix}
\alias{create_absence_presence_matrix}
\title{create absence/presence matrix for Jaccard Index calculation}
\usage{
create_absence_presence_matrix(s, c, d, ss)
}
\arguments{
\item{s}{character array}

\item{c}{character array}

\item{d}{data.frame}

\item{ss}{numeric}
}
\value{
matrix, array
}
\description{
create absence/presence matrix for Jaccard Index calculation
}
