% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dice_coeff.R
\name{dice_coeff}
\alias{dice_coeff}
\title{Sorensen-Dice Coefficient}
\usage{
dice_coeff(matrix1, matrix2)
}
\arguments{
\item{matrix1}{matrix, array}

\item{matrix2}{matrix, array}
}
\value{
numeric
}
\description{
Sorensen-Dice Coefficient
}
\examples{
\dontrun{dice_coeff(matrix1,matrix2)}
}
