% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaccard.index.R
\name{jaccard.index}
\alias{jaccard.index}
\title{Jaccard Index}
\usage{
jaccard.index(a, b)
}
\arguments{
\item{a}{matrix, array}

\item{b}{matrix, array}
}
\value{
numeric
}
\description{
This function calculates the Jaccard Index as following:
    1. Compare each cell of matrices individually. 
    2. Find the number of cells having the same content. (intersection)
    3. Calculate the total number of cells in two matrices.
    4. Subtract the intersection from the total number of cells.(union)
    5. Jaccard Index = intersection/union
The size of two matrices must be equal to calculate the Jaccard Index.
modified from: https://www.r-bloggers.com/2021/11/how-to-calculate-jaccard-similarity-in-r-2/
}
