% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transriv_statistics.R
\name{mean_transition}
\alias{mean_transition}
\title{Mean of each transition parameter}
\usage{
mean_transition(data)
}
\arguments{
\item{data}{data.frame -- the preprocessed transition data}
}
\value{
data.frame -- It contains the mean values of each measure and their
    standard deviation values.
}
\description{
Mean and standard deviation of frequency, duration and speed are calculated.
}
\examples{
\dontrun{
data <- read_csv_data(csv_path)
data <- predefine_immediate(data,0)
mean_values <- mean_transition(data)}
}
