% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transriv_statistics.R
\name{median_transition}
\alias{median_transition}
\title{Median of each transition parameter}
\usage{
median_transition(data)
}
\arguments{
\item{data}{data.frame -- the preprocessed transition data}
}
\value{
data.frame -- It contains the median values of each measure and their
    median absolute values.
}
\description{
Median and median absolute deviation of frequency, duration and speed are
calculated.
}
\examples{
\dontrun{
data <- read_csv_data(csv_path)
data <- predefine_immediate(data,0)
med_values <- median_transition(data)}
}
