% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transriv_statistics.R
\name{normalize_data}
\alias{normalize_data}
\title{Make all the small adjustments and corrections for transition data}
\usage{
normalize_data(data)
}
\arguments{
\item{data}{data.frame -- preprocessed transition data}
}
\value{
data.frame -- normalized & corrected transition data
}
\description{
Make all the small adjustments and corrections for transition data
}
\examples{
\dontrun{
data <- normalize_data(data)}
}
