% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_frequency.R
\name{normalize_frequency}
\alias{normalize_frequency}
\title{Normalize frequency in each run}
\usage{
normalize_frequency(data)
}
\arguments{
\item{data}{data.frame -- preprocessed transition data}
}
\value{
data.frame -- transition data with normalized frequency values
}
\description{
This function divides each frequency value by the sum of frequencies in the
corresponding run so that all the transition types have normalized and
comparable relative frequencies.
}
\examples{
\dontrun{
data <- read_csv_data(csv_path)
data <- predefine_immediate(data,0)
data <- normalize_frequency(data)}
}
