% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stats.R
\name{plot_stats}
\alias{plot_stats}
\title{Plot descriptive statistics with error bars}
\usage{
plot_stats(stat_data, stat_val, trans_val, cat_names, convert_na)
}
\arguments{
\item{stat_data}{data.frame -- all statistical results of transition parameters
The output of \link{descriptive_transition} can be used.}

\item{stat_val}{character -- either mean or median to extract data}

\item{trans_val}{character -- 'frequency', 'duration' or 'speed' to extract data
If not defined, all parameters will be drawn on a plot.}

\item{cat_names}{character -- category names as a character array for x-label}

\item{convert_na}{logical -- whether or not the na values in statistics are
desired to be converted into zero values}
}
\value{
ggplot
}
\description{
This function can use the output of descriptive stat functions
\link{descriptive_transition}, \link{mean_transition} or
\link{median_transition} to visualize the results.
}
\note{
The error bars are drawn as given in the stat_data. The default of this
package is standard deviation for mean and median absolute value for median.
}
\examples{
\dontrun{
plot_stats(all_stats,'median','frequency')
plot_stats(all_stats,'mean')
}

}
