% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transriv_preprocessing.R
\name{predefine_immediate}
\alias{predefine_immediate}
\title{Defining the category as immediate when there is no mixed percept}
\usage{
predefine_immediate(data, save_flag, csv_path)
}
\arguments{
\item{data}{data.frame -- the output of preprocessing steps}

\item{save_flag}{logical -- TRUE if you want to save the output as csv}

\item{csv_path}{character -- where to write the output csv file}
}
\value{
data.frame -- same data dtructure with added 'immediate' categories
}
\description{
Defining the category as immediate when there is no mixed percept
}
\examples{
\dontrun{
data <- predefine_immediate(data)
predefine_immediate(data, 1, csv_path)}
}
