% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presence_plot.R
\name{presence_plot}
\alias{presence_plot}
\title{Draw presence plot}
\usage{
presence_plot(curr_data, curr_title, upper_limit)
}
\arguments{
\item{curr_data}{data.frame}
}
\value{
ggplot
}
\description{
Draw presence plot
}
