% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transriv_preprocessing.R
\name{read_csv_data}
\alias{read_csv_data}
\title{Reading the csv file containing the questionnaire data}
\usage{
read_csv_data(filelocation)
}
\arguments{
\item{filelocation}{char -- the path of the file including the filename and
extension. The file is assumed as csv file separated by tab and decimal
point is defined as '.'.}
}
\value{
data.frame
}
\description{
Reading the csv file containing the questionnaire data
}
\note{
The function warns with a message if the file doesn't exist or cannot
    be read.
}
\examples{
\dontrun{
read_csv_data('Users/username/Documents/questionnaire.csv')}
}
