% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transriv_preprocessing.R
\name{read_mat_data}
\alias{read_mat_data}
\title{Reading the mat file containing the questionnaire data of individual subject}
\usage{
read_mat_data(filelocation)
}
\arguments{
\item{filelocation}{char -- the path of the file including the filename and
extension. The file is created on matlab and includes the questionnaires
for each run and the demographic info about the subject.}
}
\value{
list
}
\description{
Reading the mat file containing the questionnaire data of individual subject
}
\note{
The function warns with a message if the file doesn't exist or cannot
    be read.
}
\examples{
\dontrun{
read_mat_data('Users/username/Documents/s001_assessment.mat')}
}
