% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riv-parameters.R
\name{riv.parameters}
\alias{riv.parameters}
\title{Calculate several rivalry parameters for given data
This function calculates several statistical values for the rivalry data. It
is inspired by matlab function analyzeRivalry() created N. Zaretskaya (2015).}
\usage{
riv.parameters(data, subject, session, run, trial)
}
\arguments{
\item{data}{dataframe}

\item{subject}{char}

\item{session}{numeric}

\item{run}{numeric}

\item{trial}{numeric}
}
\value{
dataframe
}
\description{
Calculate several rivalry parameters for given data
This function calculates several statistical values for the rivalry data. It
is inspired by matlab function analyzeRivalry() created N. Zaretskaya (2015).
}
