% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transriv_preprocessing.R
\name{combine_all_subjects}
\alias{combine_all_subjects}
\title{Combining the csv and mat data into csv file for the given subjects}
\usage{
combine_all_subjects(
  csv_path,
  mat_folder,
  subject_list,
  basename,
  after_basename
)
}
\arguments{
\item{csv_path}{character -- the fullpath of csv file}

\item{mat_folder}{character -- the path of folder containing all the mat files}

\item{subject_list}{character -- It can be an array containing all the
subject ids such as c('s001','s002','s003')}

\item{basename}{character -- the basename of the mat files such as
'assessments_'}

\item{after_basename}{logical -- The file name of mat file is created
as paste0(basename,subject_id) if 1 and
as paste0(subject_id,basename) if 0.}
}
\value{
data.frame
}
\description{
This function adds the demographic data and the quantitative data of
transition questionnaire by using \link{combine_csv_mat} for every subject.
}
\note{
It uses \link{combine_csv_mat} function.

It does not change the file. If the file is wanted to be changed,
    one must save the output as csv.
}
\examples{
\dontrun{
data <- combine_all_subjects(csv_path,mat_folder,subject_list,'assessments_',1)
write.csv(data, file = csv_path) # to save onto the csv file}
}
