% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transriv_statistics.R
\name{descriptive_transition}
\alias{descriptive_transition}
\title{Descriptive statistics of transition parameters}
\usage{
descriptive_transition(data)
}
\arguments{
\item{data}{data.frame -- preprocessed transition data}
}
\value{
data.frame -- descriptive statistics (median, mad and mean, std)
}
\description{
Before calculating the statistics, the category without a mixed percept is
defined by using \link{transrivalry::predefine_immediate} function. Then, the
mean, standard deviation, median, median absolute deviation of frequency,
duration and speed are calculated for every transition types.
}
\examples{
\dontrun{
all_stats <- descriptive_transition(data)}
}
