% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transriv_preprocessing.R
\name{predefine_immediate}
\alias{predefine_immediate}
\title{Defining the category as immediate when there is no mixed percept}
\usage{
predefine_immediate(data, save_flag, csv_path)
}
\arguments{
\item{data}{data.frame -- the output of preprocessing steps}

\item{save_flag}{logical -- TRUE if you want to save the output as csv}

\item{csv_path}{character -- where to write the output csv file}
}
\value{
data.frame -- same data structure with added 'immediate' categories
}
\description{
This function defines the category of transition as 'immediate' with the
category code 0 when the mixed percept is absent ('no' in mixed_percept
column).
}
\note{
If save_flag and csv_path variables are defined, the function saves the
    output onto the csv file. If not, the output can be used only in the
    current session as defined variable.
}
\examples{
\dontrun{
data <- predefine_immediate(data)
predefine_immediate(data, 1, csv_path) # to save onto the csv file}
}
