% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_csv_data.R
\name{read_csv_data}
\alias{read_csv_data}
\title{Reading the csv file containing the questionnaire data}
\usage{
read_csv_data(filelocation)
}
\arguments{
\item{filelocation}{char -- the path of the file including the filename and
extension. The file is assumed as csv file separated by comma ',' or
semicomma ';' and decimal point is defined as '.'.}
}
\value{
data.frame -- the transition data
}
\description{
This functions reads the csv file containing all the questionnaire data.
}
\details{
This csv file is created by the manual examination of each questionnaire.
During this process, the transition appearances are categorized by content
analysis for the drawings and description text drawn and written by the
participants.

The demographics and the quantitative data of the questionnaire are added
later by the function \link{combine_csv_mat}.
}
\note{
The function warns with a message if the file doesn't exist or cannot
    be read.
}
\examples{
\dontrun{
read_csv_data('/Users/username/Documents/questionnaire.csv')}
}
