% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mat_data.R
\name{read_mat_data}
\alias{read_mat_data}
\title{Reading the mat file containing the questionnaire data and demographics}
\usage{
read_mat_data(filelocation)
}
\arguments{
\item{filelocation}{char -- the path of the file including the file name and
extension. The file is created on matlab and includes the questionnaires
for each run and the demographic info about the subject.}
}
\value{
list
}
\description{
This functions reads the mat files created in matlab. The mat files include
the demographic data which is collected on a survey and the quantitative data
of the transition questionnaire. Those data are combined on matlab and saved
as one mat file per subject.
}
\note{
The function warns with a message if the file doesn't exist or cannot
    be read.
}
\examples{
\dontrun{
read_mat_data('Users/username/Documents/s001_assessment.mat')}
}
